requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
sentence-transformers
faiss-cpu
pypdf
groq
